fix(RVCV-147): 삭제된 리뷰 제외

+++ domain-module/review/src/main/java/com/romanticpipe/reviewcanvas/service/ReviewService.java
@@ -83,6 +83,6 @@ public class ReviewService {
 	}
 
 	public Integer countByShopAdminId(Integer shopAdminId) {
-		return Math.toIntExact(reviewRepository.countByShopAdminId(shopAdminId));
+		return reviewRepository.countByShopAdminId(shopAdminId).intValue();
 	}
 }
